refactor(tags): migration typescript
